moved protobuf package for test inside the grpc test folder
